massive reorg. cutting down nesting
