+++ treetl/tools/polytree.py
